+++ providers/eks/standard/eks_test.go
@@ -11,7 +11,8 @@ var _ = Describe("Common tests", func() {
 	// EKS does not have metrics for k8s control plane components.
 	cfg.ControlPlaneMetricsSupported = false
 	// EKS doesn't have any of the Giant Swarm apps deployed
-	cfg.MinimalCluster = true
+	cfg.ObservabilityBundleInstalled = false
+	cfg.SecurityBundleInstalled = false
 	cfg.ExternalDnsSupported = false
 	cfg.AutoScalingSupported = false
 	common.Run(cfg)